+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # PayPal iOS SDK Release Notes
 
+## unreleased
+* CorePayments
+  * Analytics
+    * Update `component` from `ppcpmobilesdk` to `ppcpclientsdk`
+    * Send `correlation_id`, when possible, in PayPal Native Checkout analytic events
+
 ## 1.0.0 (2023-10-02)
 * Breaking Changes
   * Require Xcode 14.3+ and Swift 5.8+